distribute sub package
