+++ filter/src/main/java/criteria/CriteriaMale.java
@@ -1,14 +1,15 @@
+package criteria;
+
+import domain.Student;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class CriteriaGPAHigherThan3 implements Criteria {
-
-    public static final double standardGpa = 3.0;
+public class CriteriaMale implements Criteria {
 
     @Override
     public List<Student> meetCriteria(final List<Student> students) {
         return students.stream()
-                .filter(student -> student.isGpaHigherThan(standardGpa))
+                .filter(Student::isMale)
                 .collect(Collectors.toList());
     }
 }
