+++ filter/src/main/java/criteria/OrCriteria.java
@@ -1,3 +1,6 @@
+package criteria;
+
+import domain.Student;
 import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
