+++ filter/src/main/java/domain/Gender.java
@@ -1,3 +1,5 @@
+package domain;
+
 public enum Gender {
     MALE,
     FEMALE
